+++ src/files/infrastructure/FileJSDataverseRepository.ts
@@ -31,6 +31,7 @@ import { FileCriteria } from '../domain/models/FileCriteria'
 import { DomainFileMapper } from './mappers/DomainFileMapper'
 import { JSFileMapper } from './mappers/JSFileMapper'
 import { DatasetVersion, DatasetVersionNumber } from '../../dataset/domain/models/Dataset'
+import { DatasetJSDataverseRepository } from '@/dataset/infrastructure/repositories/DatasetJSDataverseRepository'
 import { File } from '../domain/models/File'
 import { FilePaginationInfo } from '../domain/models/FilePaginationInfo'
 import { requireAppConfig } from '../../config'
@@ -293,7 +294,11 @@ export class FileJSDataverseRepository implements FileRepository {
               FileJSDataverseRepository.getDownloadCountById(jsFile.id, jsFile.publicationDate),
               Promise.resolve(resolvedPermissions),
               FileJSDataverseRepository.getThumbnailById(jsFile.id),
-              FileJSDataverseRepository.getTabularDataById(jsFile.id, jsFile.tabularData)
+              FileJSDataverseRepository.getTabularDataById(jsFile.id, jsFile.tabularData),
+              DatasetJSDataverseRepository.isLatestReleasedVersion(
+                jsDataset.id,
+                datasetVersionNumber
+              )
             ])
           })
       })
@@ -306,7 +311,8 @@ export class FileJSDataverseRepository implements FileRepository {
           downloadsCount,
           permissions,
           thumbnail,
-          tabularData
+          tabularData,
+          isLatestDatasetVersion
         ]) =>
           JSFileMapper.toFile(
             jsFile,
@@ -316,7 +322,8 @@ export class FileJSDataverseRepository implements FileRepository {
             downloadsCount,
             permissions,
             thumbnail,
-            tabularData
+            tabularData,
+            isLatestDatasetVersion
           )
       )
       .catch((error: ReadError) => {