+++ src/files/infrastructure/mappers/JSFileMapper.ts
@@ -48,7 +48,8 @@ export class JSFileMapper {
     downloadsCount: number,
     permissions: FilePermissions,
     thumbnail?: string,
-    tabularData?: FileTabularData
+    tabularData?: FileTabularData,
+    isLatestDatasetVersion = true
   ): File {
     const datasetVersion = JSDatasetVersionMapper.toVersion(
       jsDataset.versionId,
@@ -57,7 +58,9 @@ export class JSFileMapper {
       datasetCitation,
       jsDataset.versionInfo.lastUpdateTime,
       jsDataset.publicationDate,
-      jsDataset.termsOfUse?.termsOfAccess
+      jsDataset.termsOfUse?.termsOfAccess,
+      undefined,
+      isLatestDatasetVersion
     )
     return {
       id: this.toFileId(jsFile.id),